various fixes  - packed length is encoded in 4 bytes instead of 2  - decompress can return error  - hadamard nr_bytes_to_compress_ is set to 3 instead of 2  - test added for 12_chan_32bit_1801625_samples - but data is not added to the repo

+++ lib_rspt/lib_zaxtensor/ZaxJsonParser.h
@@ -387,7 +387,7 @@ class ZaxJsonParser
         if (a_error_output)
         {
             if (errno == 0) {
-                if (tmp <= UINT16_MAX && tmp >= 0)
+                if (tmp <= (int)UINT16_MAX && tmp >= 0)
                     a_dst = tmp;
                 else
                     a_error_output->push_back(std::string("ERROR: error parsing uint16_t in JSON, out of range: '") + a_json + "'\n");
